+++ notification-center/src/Dignite.Abp.NotificationCenter.EntityFrameworkCore/Dignite/Abp/NotificationCenter/EntityFrameworkCore/INotificationCenterDbContext.cs
@@ -20,4 +20,6 @@ public interface INotificationCenterDbContext : IEfCoreDbContext
     DbSet<NotificationQuietHours> NotificationQuietHours { get; }
 
     DbSet<NotificationRetentionCleanupCursor> NotificationRetentionCleanupCursors { get; }
+
+    DbSet<NotificationAudienceBroadcastState> NotificationAudienceBroadcastStates { get; }
 }