+++ notification-center/test/Dignite.Abp.NotificationCenter.MongoDB.Tests/NotificationAudienceBroadcastProgressStore_Tests.cs
@@ -1,0 +1,10 @@
+using Dignite.Abp.NotificationCenter.MongoDB;
+using Xunit;
+
+namespace Dignite.Abp.NotificationCenter;
+
+[Collection(MongoTestCollection.Name)]
+public class NotificationAudienceBroadcastProgressStore_Tests :
+    NotificationAudienceBroadcastProgressStore_Tests<AbpNotificationCenterMongoDbTestModule>
+{
+}