+++ notification-center/test/Dignite.Abp.NotificationCenter.Tests/NotificationAudienceBroadcastProgressStore_Tests.cs
@@ -1,0 +1,6 @@
+namespace Dignite.Abp.NotificationCenter;
+
+public class NotificationAudienceBroadcastProgressStore_Tests :
+    NotificationAudienceBroadcastProgressStore_Tests<AbpNotificationCenterEntityFrameworkCoreTestModule>
+{
+}